Set default server port to 443.
Since the code now always forces HTTPS (ie., port 443) connections to
MUSICBRAINZ_SERVERS, having port set to 80 means that the interface
shows musicbrainz.org:80 as the default, while it will actually be
connecting to musicbrainz.org:443. So this change makes the interface
more consistent with what is actually going on.

See discussion from and onwards:
https://github.com/metabrainz/picard/pull/600#issuecomment-275917095